+++ src/Engine/Settings.cpp
@@ -10,16 +10,16 @@
 #include "Engine2D/Physics/Physics2D.hpp"
 
 namespace Engine {
-  glm::vec<2, size_t> Settings::Window::resolution = glm::vec<2, size_t>(800, 600);
-  std::string Settings::Window::title = "Game Window";
-  bool Settings::Window::allowResize = true;
-
   void Settings::Window::SetScreenResolution(const glm::vec<2, size_t> newValue) {
+    if (!Engine2D::Game2D::instance)
+      return;
     resolution = newValue;
     glfwSetWindowSize(Engine2D::Game2D::instance->window, newValue.x, newValue.y);
   }
 
   void Settings::Window::SetTitle(const std::string &newTitle) {
+    if (!Engine2D::Game2D::instance)
+      return;
     title = std::move(newTitle);
     glfwSetWindowTitle(Engine2D::Game2D::instance->window, title.c_str());
   }
@@ -40,10 +40,6 @@ namespace Engine {
     return allowResize;
   }
 
-  bool Settings::Graphics::vsyncEnabled = false;
-  bool Settings::Graphics::maintainAspectRatio = true;
-  unsigned int Settings::Graphics::targetFrameRate = 0;
-
   void Settings::Graphics::SetVsyncEnabled(const bool newState) {
     vsyncEnabled = newState;
     glfwSwapInterval(vsyncEnabled ? 1 : 0);
@@ -59,8 +55,10 @@ namespace Engine {
   }
 
   void Settings::Graphics::SetTargetFrameRate(const unsigned int newValue) {
-    targetFrameRate = newValue;
+    if (!Engine2D::Game2D::instance)
+      return;
 
+    targetFrameRate = newValue;
     if (targetFrameRate > 0) {
       Engine2D::Game2D::instance->targetFrameRate = targetFrameRate;
       Engine2D::Game2D::instance->targetRenderRate = 1.0f / targetFrameRate;
@@ -82,20 +80,10 @@ namespace Engine {
     return targetFrameRate;
   }
 
-  float Settings::Physics::fixedDeltaTime = 0.02f;
-  bool Settings::Physics::frictionEnabled = true;
-  bool Settings::Physics::useScreenPartitioning = false;
-  glm::vec<2, size_t> Settings::Physics::partitionSize = glm::vec<2, size_t>(4, 4);
-  glm::vec2 Settings::Physics::gravity = glm::vec2(0.0f, -9.81f);
-
   void Settings::Physics::SetFixedDeltaTime(const float newValue) {
-    constexpr float min = 1.0f / 50.0f;
-    constexpr float max = 1.0f / 60.0f;
+    static constexpr float min = 0.02f;
+    static constexpr float max = 1.0f / 60.0f;
     fixedDeltaTime = std::clamp(newValue, min, max);
-  }
-
-  void Settings::Physics::SetFrictionEnabled(const bool newState) {
-    frictionEnabled = newState;
   }
 
   void Settings::Physics::SetUseScreenPartitioning(const bool newState) {
@@ -115,10 +103,6 @@ namespace Engine {
     return fixedDeltaTime;
   }
 
-  float Settings::Physics::GetFrictionEnabled() {
-    return frictionEnabled;
-  }
-
   bool Settings::Physics::GetUseScreenPartitioning() {
     return useScreenPartitioning;
   }
@@ -130,12 +114,6 @@ namespace Engine {
   glm::vec2 Settings::Physics::GetGravity() {
     return gravity;
   }
-
-  bool Settings::Input::allowMouseInput = true;
-  bool Settings::Input::allowKeyboardInput = true;
-  bool Settings::Input::allowGamepadInput = true;
-  float Settings::Input::gamepadStickThreshold = 0.15f;
-  float Settings::Input::gamepadTriggerThreshold = -1.0f;
 
   void Settings::Input::SetAllowMouseInput(const bool newState) {
     allowMouseInput = newState;